Further refinement of the leaderboard server endpoints
Seems to be roughly functioning, but only takes in a JSON file atm

+++ server/src/models/db/leaderboard-entry.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, type) => {
 		},
 		tickrate: type.SMALLINT.UNSIGNED,
 		dateAchieved: type.DATE,
-		time: type.FLOAT.UNSIGNED,
+		time: type.DOUBLE.UNSIGNED, // TODO: investigate decimal, float and double (got weird up results with float)
 		flags: type.INTEGER.UNSIGNED,
 		file: type.STRING,
 	})
